♻️ refactor: initalstate 분리
