+++ src/contexts/AuthProvider/index.jsx
@@ -3,21 +3,9 @@ import React, { createContext, useReducer } from 'react'
 import PropTypes from 'prop-types'
 import Header from '@components/ui/Header'
 import Spinner from '@components/templates/Spinner'
+import initialState from './initalstate'
 import AuthReducer from './reducer'
 import useActions from './actions'
-
-const initialState = {
-  isLoading: false,
-  token: null,
-  userData: {
-    id: null,
-    email: null,
-    appleLevel: null,
-    nickname: null,
-    profileImage: null,
-    address: null,
-  },
-}
 
 export const AuthContext = createContext()
 
@@ -38,14 +26,28 @@ const AuthProvider = ({ children }) => {
 
   if (state.isLoading) {
     return (
-      <div className="container">
-        <div className="wrapper">
-          <Header />
-          <div className="spinner-wrapper">
-            <Spinner loading={state.isLoading} size={70} />
+      <AuthContext.Provider
+        value={{
+          state,
+          handleEmailLogin,
+          handleKakaoLogin,
+          handleSignup,
+          handleWithDrawal,
+          handleLogout,
+          handleGetUserInfo,
+          handleModifyUserInfo,
+          handleLoadingOn,
+          handleLoadingOff,
+        }}>
+        <div className="container">
+          <div className="wrapper">
+            <Header />
+            <div className="spinner-wrapper">
+              <Spinner loading={state.isLoading} size={70} />
+            </div>
           </div>
         </div>
-      </div>
+      </AuthContext.Provider>
     )
   }
