+++ src/contexts/AuthProvider/initalstate.js
@@ -1,0 +1,14 @@
+const initialState = {
+  isLoading: false,
+  token: null,
+  userData: {
+    id: null,
+    email: null,
+    appleLevel: null,
+    nickname: null,
+    profileImage: null,
+    address: null,
+  },
+}
+
+export default initialState
